Fixed few alignment issues in header and sidebar
